# backend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kapital API Documentation",
    "version": "1.0.0",
    "description": "API documentation for Kapital backend services"
  },
  "servers": [
    {
      "url": "http://localhost:4001",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "LoanApplication": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the loan application"
          },
          "userId": {
            "type": "string",
            "description": "ID of the user who created the application"
          },
          "productId": {
            "type": "string",
            "description": "ID of the selected loan product"
          },
          "amount": {
            "type": "number",
            "description": "Loan amount requested"
          },
          "term": {
            "type": "number",
            "description": "Loan term in months"
          },
          "purpose": {
            "type": "string",
            "description": "Purpose of the loan"
          },
          "monthlyRepayment": {
            "type": "number",
            "description": "Monthly repayment amount"
          },
          "interestRate": {
            "type": "number",
            "description": "Interest rate percentage"
          },
          "lateFee": {
            "type": "number",
            "description": "Late fee amount"
          },
          "originationFee": {
            "type": "number",
            "description": "Origination fee amount"
          },
          "legalFee": {
            "type": "number",
            "description": "Legal fee amount"
          },
          "applicationFee": {
            "type": "number",
            "description": "Application fee amount"
          },
          "netDisbursement": {
            "type": "number",
            "description": "Net amount to be disbursed after fees"
          },
          "urlLink": {
            "type": "string",
            "description": "Unique URL link for the application"
          },
          "status": {
            "type": "string",
            "enum": [
              "INCOMPLETE",
              "PENDING",
              "APPROVED",
              "REJECTED",
              "DISBURSED",
              "CLOSED"
            ],
            "description": "Current status of the application"
          },
          "appStep": {
            "type": "number",
            "description": "Current step in the application process (0-4)"
          },
          "acceptTerms": {
            "type": "boolean",
            "description": "Whether the user has accepted the terms"
          },
          "paidAppFee": {
            "type": "boolean",
            "description": "Whether the application fee has been paid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the application was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the application was last updated"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Document"
            }
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the product"
          },
          "code": {
            "type": "string",
            "description": "Product code"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "description": {
            "type": "string",
            "description": "Product description"
          },
          "minAmount": {
            "type": "number",
            "description": "Minimum loan amount"
          },
          "maxAmount": {
            "type": "number",
            "description": "Maximum loan amount"
          },
          "repaymentTerms": {
            "type": "array",
            "items": {
              "type": "number"
            },
            "description": "Available repayment terms in months"
          },
          "interestRate": {
            "type": "number",
            "description": "Interest rate percentage"
          },
          "eligibility": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Eligibility criteria"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the product is active"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the product was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the product was last updated"
          }
        }
      },
      "ProductInput": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Product code"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "description": {
            "type": "string",
            "description": "Product description"
          },
          "minAmount": {
            "type": "number",
            "description": "Minimum loan amount"
          },
          "maxAmount": {
            "type": "number",
            "description": "Maximum loan amount"
          },
          "repaymentTerms": {
            "type": "array",
            "items": {
              "type": "number"
            },
            "description": "Available repayment terms in months"
          },
          "interestRate": {
            "type": "number",
            "description": "Interest rate percentage"
          },
          "eligibility": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Eligibility criteria"
          },
          "lateFee": {
            "type": "number",
            "description": "Late payment fee percentage"
          },
          "originationFee": {
            "type": "number",
            "description": "Origination fee percentage"
          },
          "legalFee": {
            "type": "number",
            "description": "Legal fee amount"
          },
          "applicationFee": {
            "type": "number",
            "description": "Application fee amount"
          },
          "requiredDocuments": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Required document types"
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Product features"
          },
          "loanTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Available loan purpose types"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the product is active"
          }
        }
      },
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the document"
          },
          "loanApplicationId": {
            "type": "string",
            "description": "ID of the loan application this document belongs to"
          },
          "documentType": {
            "type": "string",
            "description": "Type of document"
          },
          "documentUrl": {
            "type": "string",
            "description": "URL to the document"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ],
            "description": "Status of the document verification"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the document was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the document was last updated"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "summary": "Get all users (admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied, admin privileges required"
          },
          "500": {
            "description": "Server error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of users to return"
          }
        ]
      }
    },
...
    "/api/admin/loans": {
      "get": {
        "summary": "Get all approved and disbursed loan applications",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved loans",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LoanApplication"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, token is invalid or missing"
          },
          "403": {
            "description": "Forbidden, user is not an admin"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/admin/notifications": {
      "get": {
        "summary": "Get all notifications (admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "SYSTEM",
                "MARKETING"
              ]
            },
            "description": "Filter by notification type"
          },
          {
            "in": "query",
            "name": "priority",
            "schema": {
              "type": "string",
              "enum": [
                "LOW",
                "MEDIUM",
                "HIGH"
              ]
            },
            "description": "Filter by priority"
          },
          {
            "in": "query",
            "name": "isRead",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by read status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of all notifications"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied, admin privileges required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/loan-applications": {
      "post": {
        "summary": "Create a new loan application",
        "tags": [
          "Loan Applications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "productId",
                  "amount",
                  "term",
                  "purpose"
                ],
                "properties": {
                  "productId": {
                    "type": "string",
                    "description": "ID of the selected loan product"
                  },
                  "amount": {
                    "type": "number",
                    "description": "Loan amount requested"
                  },
                  "term": {
                    "type": "number",
                    "description": "Loan term in months"
                  },
                  "purpose": {
                    "type": "string",
                    "description": "Purpose of the loan"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Loan application created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanApplication"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "get": {
        "summary": "Get all loan applications for the current user",
        "tags": [
          "Loan Applications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of loan applications retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoanApplication"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/loans": {
      "get": {
        "summary": "Get user's loans",
        "tags": [
          "Loans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user's loans (active, pending discharge, and discharged)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "loans": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "principalAmount": {
                            "type": "number"
                          },
                          "outstandingBalance": {
                            "type": "number"
                          },
                          "interestRate": {
                            "type": "number"
                          },
                          "term": {
                            "type": "integer"
                          },
                          "monthlyPayment": {
                            "type": "number"
                          },
                          "nextPaymentDue": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "status": {
                            "type": "string"
                          },
                          "disbursedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "application": {
                            "type": "object",
                            "properties": {
                              "product": {
                                "type": "object",
                                "properties": {
                                  "name": {
                                    "type": "string"
                                  },
                                  "code": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "summary": "Get user's notifications",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of notifications"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "patch": {
        "summary": "Mark notifications as read",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "notificationIds"
                ],
                "properties": {
                  "notificationIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notifications marked as read"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/products": {
      "get": {
        "summary": "Get all active products or a specific product by code",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "schema": {
              "type": "string"
            },
            "description": "Product code to filter by"
          }
        ],
        "responses": {
          "200": {
            "description": "List of active products or a single product if code is provided",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create a new product",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}